Prelim PvP stuff
I'm sure this will be controversial, but I think this will be good for
ShadowCraft on the whole. It will let players use Shadowcraft to
measure the value of PvP Power gems. Also, if we ever get into
measuring burst scenarios it could be immensely helpful. TODO: target
armor, rating conversion to %s.